Macro time series data euro area.
The data corresponds to the 19 countries of the Euro Area. GDP, GDP implicit price deflator,and potential GDP are produced by Eurostat. The interbank rate is produced by the OECD.The potential GDP was download from AMECO database. The GDP, the GDP implicit pricedeflator, and the interbank rate were downloaded from DBnomics.